unification of scripts naming
