+++ src/generators/__init__.py
@@ -1,1 +1,1 @@
-"""Runners of the experiments."""
+"""Runners of the experiments to generate dataset."""
