+++ src/generators/commons.py
@@ -1,30 +1,9 @@
 import warnings
-from dataclasses import dataclass
-from itertools import product
 from math import log10
 
 import network_diffusion as nd
-from _data_set.nsl_data_utils.loaders.net_loader import load_network
 
 warnings.filterwarnings(action="ignore", category=FutureWarning)
-
-
-@dataclass(frozen=True)
-class Network:
-    name: str
-    type: str # this field has been added after generation of the dataset with spreading potentials!
-    graph: nd.MultilayerNetwork | nd.MultilayerNetworkTorch
-
-
-def get_parameter_space(
-    protocols: list[str], p_values: list[float], networks: list[str], as_tensor: bool
-) -> product:
-    nets = []
-    for net_type in networks:
-        print(f"Loading {net_type} network")
-        for net_name, net_graph in load_network(net_name=net_type, as_tensor=as_tensor).items():
-            nets.append(Network(name=net_name, type=net_type, graph=net_graph))
-    return product(protocols, p_values, nets)
 
 
 def get_ranking(actor: nd.MLNetworkActor, actors: list[nd.MLNetworkActor]) -> nd.seeding.MockingActorSelector:
