+++ src/generators/step_tensor.py
@@ -1,4 +1,4 @@
-"""Main runner of the simulator."""
+"""Single simulation step implemented with `torch`."""
 
 from pathlib import Path
 
@@ -7,11 +7,8 @@ from tqdm import tqdm
 
 from src.icm.torch_model import TorchMICModel, TorchMICSimulator
 from src.generators import commons
-from src.generators.utils import (
-    mean_simulation_results,
-    save_magrinal_efficiences,
-    SimulationResult,
-)
+from src.generators.utils import mean_simulation_results, SimulationResult
+from src.utils import export_dataclasses
 
 
 def experiment_step(
@@ -53,7 +50,12 @@ def experiment_step(
             )
 
             # run experiment on a deep copy of the network!
-            experiment = TorchMICSimulator(model=micm, net=net, n_steps=len(net.actors_map) * 2, seed_set={actor_name})
+            experiment = TorchMICSimulator(
+                model=micm,
+                net=net,
+                n_steps=len(net.actors_map) * 2,
+                seed_set={actor_name},
+            )
             logs = experiment.perform_propagation()
 
             # compute boost that current actor provides
@@ -75,4 +77,4 @@ def experiment_step(
 
     # save efficiences obtained for this case
     investigated_case_file_path = out_dir / f"{commons.get_case_name_base(protocol, p, net_name)}.csv"
-    save_magrinal_efficiences(marginal_efficiencies, investigated_case_file_path)
+    export_dataclasses(marginal_efficiencies, investigated_case_file_path)
